+++ apps/chat/src/store/import-export/importExport.epics.ts
@@ -31,11 +31,15 @@ import {
 import { BucketService } from '@/src/utils/app/data/bucket-service';
 import { ConversationService } from '@/src/utils/app/data/conversation-service';
 import { FileService } from '@/src/utils/app/data/file-service';
-import { PromptService } from '@/src/utils/app/data/prompt-service';
+import {
+  PromptService,
+  getImportPreparedPrompts,
+} from '@/src/utils/app/data/prompt-service';
 import {
   getImportPreparedConversations,
   getOrUploadConversation,
 } from '@/src/utils/app/data/storages/api/conversation-api-storage';
+import { getOrUploadPrompt } from '@/src/utils/app/data/storages/api/prompt-api-storage';
 import { BrowserStorage } from '@/src/utils/app/data/storages/browser-storage';
 import { constructPath } from '@/src/utils/app/file';
 import {
@@ -50,7 +54,9 @@ import {
   cleanPromptsFolders,
   exportConversation,
   exportConversations,
+  exportPrompt,
   exportPrompts,
+  isPromptsFormat,
   updateAttachment,
 } from '@/src/utils/app/import-export';
 import { translate } from '@/src/utils/app/translation';
@@ -99,24 +105,23 @@ const exportConversationEpic: AppEpic = (action$, state$) =>
     filter(ImportExportActions.exportConversation.match),
     switchMap(({ payload }) =>
       forkJoin({
-        conversationAndPayload: getOrUploadConversation(
+        conversation: getOrUploadConversation(
           { id: payload.conversationId },
           state$.value,
-        ),
+        ).pipe(map((data) => data.conversation)),
         withAttachments: of(payload.withAttachments),
         bucket: BucketService.getBucket(),
       }),
     ),
-    switchMap(({ conversationAndPayload, withAttachments, bucket }) => {
-      const { conversation } = conversationAndPayload;
+    switchMap(({ conversation, withAttachments, bucket }) => {
       if (!conversation) {
         return of(ImportExportActions.exportFail());
       }
+
       const parentFolders = ConversationsSelectors.selectParentFolders(
         state$.value,
         conversation.folderId,
       );
-
       const appName = SettingsSelectors.selectAppName(state$.value);
 
       if (!withAttachments) {
@@ -128,6 +133,7 @@ const exportConversationEpic: AppEpic = (action$, state$) =>
       if (!bucket.length) {
         return of(ImportExportActions.exportFail());
       }
+
       const attachments = ConversationsSelectors.getAttachments(
         state$.value,
         conversation.id,
@@ -211,6 +217,85 @@ const exportConversationsEpic: AppEpic = (action$, state$) =>
         of(
           UIActions.showErrorToast(
             translate('An error occurred while uploading conversations'),
+          ),
+        ),
+        of(ImportExportActions.exportFail()),
+      ),
+    ),
+  );
+
+const exportPromptEpic: AppEpic = (action$, state$) =>
+  action$.pipe(
+    filter(ImportExportActions.exportPrompt.match),
+    switchMap(({ payload }) => getOrUploadPrompt(payload, state$.value)),
+
+    switchMap((promptAndPayload) => {
+      const { prompt } = promptAndPayload;
+      if (!prompt) {
+        return concat(
+          of(
+            UIActions.showErrorToast(
+              translate('An error occurred while uploading prompt'),
+            ),
+          ),
+          of(ImportExportActions.exportFail()),
+        );
+      }
+
+      const appName = SettingsSelectors.selectAppName(state$.value);
+
+      exportPrompt(
+        prompt,
+        PromptsSelectors.selectParentFolders(state$.value, prompt.folderId),
+        appName,
+      );
+      return EMPTY;
+    }),
+  );
+
+const exportPromptsEpic: AppEpic = (action$, state$) =>
+  action$.pipe(
+    filter(ImportExportActions.exportPrompts.match),
+    switchMap(() =>
+      //listing of all entities
+      PromptService.getPrompts(undefined, true),
+    ),
+    switchMap((promptsListing) => {
+      const onlyMyPromptsListing = filterOnlyMyEntities(promptsListing);
+      const foldersIds = uniq(
+        onlyMyPromptsListing.map((info) => info.folderId),
+      );
+      //calculate all folders;
+      const foldersWithPrompts = getFoldersFromIds(
+        uniq(foldersIds.flatMap((id) => getParentFolderIdsFromFolderId(id))),
+        FolderType.Prompt,
+      );
+
+      const allFolders = PromptsSelectors.selectFolders(state$.value);
+
+      const folders = combineEntities(foldersWithPrompts, allFolders);
+
+      return forkJoin({
+        //get all prompts from api
+        prompts: zip(
+          onlyMyPromptsListing.map((info) => PromptService.getPrompt(info)),
+        ),
+        folders: of(folders),
+      });
+    }),
+    switchMap(({ prompts, folders }) => {
+      const filteredPrompts = prompts.filter(Boolean) as Prompt[];
+
+      const appName = SettingsSelectors.selectAppName(state$.value);
+
+      exportPrompts(filteredPrompts, folders, appName);
+      return EMPTY;
+    }),
+    catchError(() =>
+      concat(
+        of(
+          UIActions.showErrorToast(
+            translate('An error occurred while uploading prompts'),
           ),
         ),
         of(ImportExportActions.exportFail()),
@@ -337,6 +422,91 @@ const importConversationsEpic: AppEpic = (action$) =>
       );
     }),
     catchError(() => of(ImportExportActions.importFail())),
+  );
+
+const importPromptsEpic: AppEpic = (action$) =>
+  action$.pipe(
+    filter(ImportExportActions.importPrompts.match),
+    switchMap(({ payload }) => {
+      const { promptsHistory } = payload;
+
+      if (!isPromptsFormat(promptsHistory)) {
+        return concat(
+          of(ImportExportActions.importPromptsFail()),
+          of(
+            UIActions.showErrorToast(
+              translate(errorsMessages.unsupportedDataFormat, {
+                ns: 'common',
+              }),
+            ),
+          ),
+        );
+      }
+
+      const preparedPrompts: Prompt[] = getImportPreparedPrompts({
+        prompts: promptsHistory.prompts,
+        folders: promptsHistory.folders,
+      });
+
+      if (!preparedPrompts.length) {
+        return of(ImportExportActions.importPromptsFail());
+      }
+
+      return PromptService.getPrompts(undefined, true).pipe(
+        switchMap((promptsListing) => {
+          const existedImportNamesPrompts = preparedPrompts.filter(
+            (importPrompt) =>
+              !isImportEntityNameOnSameLevelUnique({
+                entity: importPrompt,
+                entities: promptsListing,
+              }),
+          );
+
+          const nonExistedImportNamesPrompts = preparedPrompts.filter(
+            (importPrompt) => {
+              return isImportEntityNameOnSameLevelUnique({
+                entity: importPrompt,
+                entities: promptsListing,
+              });
+            },
+          );
+
+          if (!existedImportNamesPrompts.length) {
+            return of(
+              ImportExportActions.uploadImportedPrompts({
+                itemsToUpload: nonExistedImportNamesPrompts,
+                folders: promptsHistory.folders,
+              }),
+            );
+          }
+
+          if (!nonExistedImportNamesPrompts.length) {
+            return of(
+              ImportExportActions.showReplaceDialog({
+                duplicatedItems: existedImportNamesPrompts,
+                featureType: FeatureType.Prompt,
+              }),
+            );
+          }
+
+          return concat(
+            of(
+              ImportExportActions.showReplaceDialog({
+                duplicatedItems: existedImportNamesPrompts,
+                featureType: FeatureType.Prompt,
+              }),
+            ),
+            of(
+              ImportExportActions.uploadImportedPrompts({
+                itemsToUpload: nonExistedImportNamesPrompts,
+                folders: promptsHistory.folders,
+              }),
+            ),
+          );
+        }),
+        catchError(() => of(ImportExportActions.importPromptsFail())),
+      );
+    }),
   );
 
 const handleDuplicatedItemsEpic: AppEpic = (action$) =>
@@ -1148,7 +1318,10 @@ const resetStateEpic: AppEpic = (action$) =>
 export const ImportExportEpics = combineEpics(
   exportConversationEpic,
   exportConversationsEpic,
+  exportPromptEpic,
+  exportPromptsEpic,
   importConversationsEpic,
+  importPromptsEpic,
   importZipEpic,
   uploadConversationAttachmentsEpic,
   uploadAllAttachmentsSuccessEpic,
